G4 offset
Set what as far as I can see is the correct offset for the G4. However, due to case differences, wear and generally me having no idea how to reasonably measure the exact perfect amount of pixels needed this might not be correct for you. I have tested and Bigboot's original DPI scaling solution works perfectly still with this change, so should work on all DPI's. If this is wrong, even by a couple of pixels and you've got an eye for that, let me know and I'll change it, rebuild and release.